Split SwiftUI section files
Move extracted child views into focused sibling files so parent screens keep orchestration code while large section groups stay easier to navigate.
