+++ ios/Cove/Flows/SettingsFlow/CloudBackupDetailSections.swift
@@ -1,12 +1,5 @@
 import SwiftUI
 
-private extension CloudOnlyOperation {
-    var operatingRecordId: String? {
-        if case let .operating(recordId) = self { return recordId }
-        return nil
-    }
-}
-
 struct DetailFormContent: View {
     let detail: CloudBackupDetail
     let syncHealth: CloudSyncHealth
@@ -221,657 +214,5 @@ struct DisableCloudBackupSection: View {
         } message: {
             Text("Disabling Cloud Backup will permanently delete your current Cove cloud backups from cloud storage. Wallets already on this device will stay on this device, but they will no longer be backed up to cloud storage.")
         }
-    }
-}
-
-struct HeaderSection: View {
-    let lastSync: UInt64?
-    let syncHealth: CloudSyncHealth
-
-    var body: some View {
-        Section {
-            VStack(spacing: 8) {
-                CloudBackupHeaderIcon(syncHealth: syncHealth)
-                    .font(.largeTitle)
-
-                Text(cloudBackupDetailHeaderTitle(syncHealth: syncHealth))
-                    .fontWeight(.semibold)
-
-                if let lastSync {
-                    Text("Last synced \(formatDate(lastSync))")
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-                }
-
-                CloudBackupSyncHealthLabel(syncHealth: syncHealth)
-            }
-            .frame(maxWidth: .infinity)
-            .padding(.vertical, 8)
-        }
-    }
-
-    private func formatDate(_ timestamp: UInt64) -> String {
-        let date = Date(timeIntervalSince1970: TimeInterval(timestamp))
-        return date.formatted(date: .abbreviated, time: .shortened)
-    }
-}
-
-private struct CloudBackupHeaderIcon: View {
-    let syncHealth: CloudSyncHealth
-
-    var body: some View {
-        let image = Image(systemName: cloudBackupDetailHeaderIconName(syncHealth: syncHealth))
-
-        switch syncHealth {
-        case .unknown:
-            image.foregroundColor(.secondary)
-        case .allUploaded:
-            image
-                .foregroundColor(.statusSuccess)
-        case .uploading:
-            image
-                .foregroundColor(.statusInfo)
-        case .failed:
-            image
-                .foregroundColor(.statusError)
-        case .authorizationRequired, .noFiles, .unavailable:
-            image
-                .foregroundColor(.statusWarning)
-        }
-    }
-}
-
-private struct CloudBackupSyncHealthLabel: View {
-    let syncHealth: CloudSyncHealth
-
-    var body: some View {
-        switch syncHealth {
-        case .unknown:
-            CloudBackupSyncProgressLabel(title: "Checking iCloud sync status...")
-        case .allUploaded:
-            Label("All files synced to iCloud", systemImage: "checkmark.circle.fill")
-                .font(.caption)
-                .foregroundStyle(Color.statusSuccess)
-        case .uploading:
-            CloudBackupSyncProgressLabel(title: "Syncing to iCloud...")
-        case let .failed(message):
-            Label("Sync error: \(message)", systemImage: "exclamationmark.triangle.fill")
-                .font(.caption)
-                .foregroundStyle(Color.statusError)
-        case .authorizationRequired:
-            Label("iCloud Drive access needs to be reconnected", systemImage: "exclamationmark.triangle.fill")
-                .font(.caption)
-                .foregroundStyle(Color.statusWarning)
-        case .noFiles:
-            Label("No iCloud backup files uploaded yet", systemImage: "icloud.slash")
-                .font(.caption)
-                .foregroundStyle(Color.statusWarning)
-        case .unavailable:
-            Label("iCloud Drive is unavailable", systemImage: "exclamationmark.triangle.fill")
-                .font(.caption)
-                .foregroundStyle(Color.statusWarning)
-        }
-    }
-}
-
-private struct CloudBackupSyncProgressLabel: View {
-    let title: String
-
-    var body: some View {
-        HStack(spacing: 4) {
-            ProgressView()
-                .controlSize(.mini)
-            Text(title)
-        }
-        .font(.caption)
-        .foregroundStyle(.secondary)
-    }
-}
-
-func cloudBackupDetailHeaderTitle(syncHealth: CloudSyncHealth) -> String {
-    switch syncHealth {
-    case .allUploaded:
-        "Cloud Backup Active"
-    case .uploading:
-        "Cloud Backup Syncing"
-    case .unknown:
-        "Checking Cloud Backup"
-    case .noFiles, .authorizationRequired, .unavailable, .failed:
-        "Cloud Backup Needs Attention"
-    }
-}
-
-func cloudBackupDetailHeaderIconName(syncHealth: CloudSyncHealth) -> String {
-    switch syncHealth {
-    case .allUploaded:
-        "checkmark.icloud.fill"
-    case .uploading:
-        "arrow.clockwise.icloud.fill"
-    case .unknown:
-        "icloud"
-    case .noFiles:
-        "icloud.slash"
-    case .authorizationRequired, .unavailable, .failed:
-        "exclamationmark.icloud.fill"
-    }
-}
-
-struct CloudOnlySection: View {
-    let manager: CloudBackupManager
-    @State private var selectedWallet: CloudBackupWalletItem?
-    @State private var walletToDelete: CloudBackupWalletItem?
-    @State private var unsupportedRestoreWallet: CloudBackupWalletItem?
-
-    private var isOperating: Bool {
-        manager.cloudOnlyOperation.operatingRecordId != nil
-    }
-
-    var body: some View {
-        Section(header: Text("Not on This Device")) {
-            CloudOnlySectionContent(
-                manager: manager,
-                isOperating: isOperating,
-                onSelectWallet: { selectedWallet = $0 }
-            )
-        }
-        .modifier(
-            CloudOnlyActionDialogs(
-                manager: manager,
-                selectedWallet: $selectedWallet,
-                walletToDelete: $walletToDelete,
-                unsupportedRestoreWallet: $unsupportedRestoreWallet
-            )
-        )
-    }
-}
-
-struct OtherBackupsSection: View {
-    let summary: CloudBackupOtherBackupsSummary
-    let manager: CloudBackupManager
-    @State private var showingRecoverConfirmation = false
-    @State private var showingDeleteConfirmation = false
-    @State private var showingFinalDeleteConfirmation = false
-    @State private var recoveryResult: OtherBackupsRecoveryResult?
-
-    private var isRecovering: Bool {
-        if case .recovering = manager.otherBackupsOperation { return true }
-        return false
-    }
-
-    private var isDeleting: Bool {
-        if case .deleting = manager.otherBackupsOperation { return true }
-        return false
-    }
-
-    private var isOperating: Bool {
-        isRecovering || isDeleting
-    }
-
-    private var failure: String? {
-        if case let .failed(error) = manager.otherBackupsOperation { return error }
-        return nil
-    }
-
-    var body: some View {
-        Section(header: Text("Other Cloud Backups")) {
-            Text(summaryText)
-                .font(.caption)
-                .foregroundStyle(.secondary)
-
-            Button {
-                showingRecoverConfirmation = true
-            } label: {
-                operationLabel(
-                    title: isRecovering ? "Trying Passkey..." : "Try Another Passkey",
-                    systemImage: "person.badge.key",
-                    isLoading: isRecovering
-                )
-            }
-            .disabled(isOperating)
-
-            Button(role: .destructive) {
-                showingDeleteConfirmation = true
-            } label: {
-                operationLabel(
-                    title: isDeleting ? "Deleting..." : "Delete These Backups",
-                    systemImage: "trash",
-                    isLoading: isDeleting
-                )
-            }
-            .disabled(isOperating)
-
-            if let failure {
-                Text(failure)
-                    .font(.caption)
-                    .foregroundStyle(Color.statusError)
-            }
-        }
-        .confirmationDialog(
-            "Recover wallets from another passkey?",
-            isPresented: $showingRecoverConfirmation,
-            titleVisibility: .visible
-        ) {
-            Button("Try Passkey") {
-                manager.dispatch(action: .recoverOtherBackups)
-            }
-            Button("Cancel", role: .cancel) {}
-        } message: {
-            Text(
-                "This will use the selected passkey once to decrypt these other backups. Your current Cloud Backup passkey will not change."
-            )
-        }
-        .alert(
-            "Wallets Recovered",
-            isPresented: Binding(
-                get: { recoveryResult != nil },
-                set: { if !$0 { recoveryResult = nil } }
-            )
-        ) {
-            Button("Verify Current Passkey") {
-                manager.startVerification(source: .cloudBackupDetail)
-            }
-            Button("Done", role: .cancel) {}
-        } message: {
-            Text(recoveryResult?.message ?? "")
-        }
-        .alert("Delete Other Cloud Backups?", isPresented: $showingDeleteConfirmation) {
-            Button("Continue", role: .destructive) {
-                showingFinalDeleteConfirmation = true
-            }
-            Button("Cancel", role: .cancel) {}
-        } message: {
-            Text("This will permanently remove these other backups from iCloud.")
-        }
-        .alert("This Cannot Be Undone", isPresented: $showingFinalDeleteConfirmation) {
-            Button("Delete", role: .destructive) {
-                manager.dispatch(action: .deleteOtherBackups)
-            }
-            Button("Cancel", role: .cancel) {}
-        } message: {
-            Text(
-                "These backups cannot be recovered later, even if you find the passkey that currently protects them."
-            )
-        }
-        .onChange(of: manager.otherBackupsOperation) { _, operation in
-            if case let .recovered(walletsRestored, walletsFailed, failedWalletErrors) = operation {
-                recoveryResult = OtherBackupsRecoveryResult(
-                    walletsRestored: walletsRestored,
-                    walletsFailed: walletsFailed,
-                    failedWalletErrors: failedWalletErrors
-                )
-            }
-        }
-    }
-
-    private var summaryText: String {
-        let namespaceLabel = pluralize(Int(summary.namespaceCount), singular: "backup set", plural: "backup sets")
-        let walletLabel = pluralize(Int(summary.walletCount), singular: "wallet", plural: "wallets")
-        let passkeyLabel = otherPasskeyLabel
-        return "\(namespaceLabel) protected by \(passkeyLabel), containing \(walletLabel)"
-    }
-
-    private var otherPasskeyLabel: String {
-        let suffixes = summary.passkeyHints.map(\.nameSuffix)
-
-        guard !suffixes.isEmpty else {
-            return "a different passkey"
-        }
-
-        if suffixes.count == 1 {
-            return "Cove Cloud Backup (\(suffixes[0]))"
-        }
-
-        return "passkeys \(suffixes.map { "(\($0))" }.joined(separator: ", "))"
-    }
-
-    private func operationLabel(title: String, systemImage: String, isLoading: Bool) -> some View {
-        HStack {
-            if isLoading {
-                ProgressView()
-                    .padding(.trailing, 4)
-            } else {
-                Image(systemName: systemImage)
-            }
-            Text(title)
-        }
-    }
-}
-
-struct OtherBackupsLoadFailedSection: View {
-    let error: String
-
-    var body: some View {
-        Section(header: Text("Other Cloud Backups")) {
-            Text("Could not load other cloud backups.")
-                .font(.caption)
-                .foregroundStyle(.secondary)
-
-            Text(error)
-                .font(.caption)
-                .foregroundStyle(Color.statusError)
-        }
-    }
-}
-
-private struct OtherBackupsRecoveryResult: Identifiable {
-    let id = UUID()
-    let walletsRestored: UInt32
-    let walletsFailed: UInt32
-    let failedWalletErrors: [String]
-
-    var message: String {
-        var parts = [
-            "Recovered \(pluralize(Int(walletsRestored), singular: "wallet", plural: "wallets")).",
-            "Your current Cloud Backup passkey is unchanged. Verify your current passkey to make sure it opens your active backup.",
-        ]
-
-        if walletsFailed > 0 {
-            parts.append(
-                "\(pluralize(Int(walletsFailed), singular: "wallet", plural: "wallets")) could not be recovered."
-            )
-        }
-
-        if let firstError = failedWalletErrors.first {
-            parts.append(firstError)
-        }
-
-        return parts.joined(separator: " ")
-    }
-}
-
-private func pluralize(_ count: Int, singular: String, plural: String) -> String {
-    "\(count) \(count == 1 ? singular : plural)"
-}
-
-private struct CloudOnlySectionContent: View {
-    let manager: CloudBackupManager
-    let isOperating: Bool
-    let onSelectWallet: (CloudBackupWalletItem) -> Void
-
-    var body: some View {
-        switch manager.cloudOnly {
-        case .notFetched, .loading:
-            HStack {
-                ProgressView()
-                    .padding(.trailing, 8)
-                Text("Loading...")
-            }
-            .foregroundStyle(.secondary)
-            .task {
-                manager.dispatch(action: .fetchCloudOnly)
-            }
-
-        case let .loaded(wallets):
-            CloudOnlyWalletRows(
-                wallets: wallets,
-                operatingRecordId: manager.cloudOnlyOperation.operatingRecordId,
-                isOperating: isOperating,
-                onSelectWallet: onSelectWallet
-            )
-
-            if case let .failed(error) = manager.cloudOnlyOperation {
-                Text(error)
-                    .font(.caption)
-                    .foregroundStyle(Color.statusError)
-            } else if case let .warning(message: message, error: _) = manager.cloudOnlyOperation {
-                Text(message)
-                    .font(.caption)
-                    .foregroundStyle(Color.statusWarning)
-            }
-
-        case let .failed(error):
-            Text(error)
-                .font(.caption)
-                .foregroundStyle(Color.statusError)
-        }
-    }
-}
-
-private struct CloudOnlyWalletRows: View {
-    let wallets: [CloudBackupWalletItem]
-    let operatingRecordId: String?
-    let isOperating: Bool
-    let onSelectWallet: (CloudBackupWalletItem) -> Void
-
-    var body: some View {
-        ForEach(wallets, id: \.recordId) { item in
-            Button {
-                onSelectWallet(item)
-            } label: {
-                HStack {
-                    if operatingRecordId == item.recordId {
-                        ProgressView()
-                            .padding(.trailing, 8)
-                    }
-                    WalletItemRow(item: item)
-                }
-            }
-            .foregroundStyle(.primary)
-            .disabled(isOperating)
-        }
-    }
-}
-
-private struct CloudOnlyActionDialogs: ViewModifier {
-    let manager: CloudBackupManager
-    @Binding var selectedWallet: CloudBackupWalletItem?
-    @Binding var walletToDelete: CloudBackupWalletItem?
-    @Binding var unsupportedRestoreWallet: CloudBackupWalletItem?
-
-    func body(content: Content) -> some View {
-        content
-            .confirmationDialog(
-                selectedWallet?.name ?? "Wallet",
-                isPresented: Binding(
-                    get: { selectedWallet != nil },
-                    set: { if !$0 { selectedWallet = nil } }
-                ),
-                titleVisibility: .visible
-            ) {
-                if let item = selectedWallet {
-                    Button("Restore to This Device") {
-                        if item.syncStatus == .unsupportedVersion {
-                            unsupportedRestoreWallet = item
-                            return
-                        }
-
-                        manager.dispatch(action: .restoreCloudWallet(item.recordId))
-                    }
-                    Button("Delete from iCloud", role: .destructive) {
-                        walletToDelete = item
-                    }
-                }
-                Button("Cancel", role: .cancel) {}
-            }
-            .alert(
-                "Delete \(walletToDelete?.name ?? "wallet")?",
-                isPresented: Binding(
-                    get: { walletToDelete != nil },
-                    set: { if !$0 { walletToDelete = nil } }
-                )
-            ) {
-                if let item = walletToDelete {
-                    Button("Delete Forever", role: .destructive) {
-                        manager.dispatch(action: .deleteCloudWallet(item.recordId))
-                    }
-                }
-                Button("Cancel", role: .cancel) {}
-            } message: {
-                Text("This wallet backup will be permanently removed from iCloud")
-            }
-            .alert(
-                "Can't Restore \(unsupportedRestoreWallet?.name ?? "Wallet")",
-                isPresented: Binding(
-                    get: { unsupportedRestoreWallet != nil },
-                    set: { if !$0 { unsupportedRestoreWallet = nil } }
-                )
-            ) {
-                Button("OK", role: .cancel) {}
-            } message: {
-                Text(
-                    "This backup uses a newer version of Cove and can't be restored on this device yet"
-                )
-            }
-    }
-}
-
-struct WalletSections: View {
-    let wallets: [CloudBackupWalletItem]
-
-    private var groupedWallets: [GroupedWalletSections.Section] {
-        GroupedWalletSections(wallets: wallets).sections
-    }
-
-    var body: some View {
-        ForEach(groupedWallets) { group in
-            Section(header: sectionHeader(for: group.key)) {
-                ForEach(group.items, id: \.recordId) { item in
-                    WalletItemRow(item: item)
-                }
-            }
-        }
-    }
-
-    private func sectionHeader(for key: GroupKey) -> some View {
-        Text(key.title)
-    }
-}
-
-private struct GroupedWalletSections {
-    struct Section: Identifiable {
-        let key: GroupKey
-        let items: [CloudBackupWalletItem]
-
-        var id: GroupKey {
-            key
-        }
-    }
-
-    let sections: [Section]
-
-    init(wallets: [CloudBackupWalletItem]) {
-        sections = Dictionary(grouping: wallets) {
-            GroupKey(network: $0.network, walletMode: $0.walletMode)
-        }
-        .map { key, items in
-            Section(key: key, items: items)
-        }
-        .sorted { $0.key < $1.key }
-    }
-}
-
-struct WalletItemRow: View {
-    let item: CloudBackupWalletItem
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            HStack {
-                Text(item.name)
-                    .fontWeight(.medium)
-                Spacer()
-                StatusBadge(status: item.syncStatus)
-            }
-
-            HStack(spacing: 12) {
-                if let network = item.network {
-                    IconLabel("globe", network.displayName())
-                }
-                if let walletType = item.walletType {
-                    IconLabel("wallet.bifold", walletType.displayName())
-                }
-                if let fingerprint = item.fingerprint {
-                    IconLabel("touchid", fingerprint)
-                }
-            }
-            .font(.caption)
-            .foregroundStyle(.secondary)
-
-            HStack(spacing: 12) {
-                if let labelCount = item.labelCount {
-                    IconLabel("tag", "\(labelCount) labels")
-                }
-                if let backupUpdatedAt = item.backupUpdatedAt {
-                    IconLabel("clock", formatDate(backupUpdatedAt))
-                }
-            }
-            .font(.caption)
-            .foregroundStyle(.secondary)
-        }
-        .padding(.vertical, 2)
-    }
-
-    private func formatDate(_ timestamp: UInt64) -> String {
-        let date = Date(timeIntervalSince1970: TimeInterval(timestamp))
-        return date.formatted(date: .abbreviated, time: .shortened)
-    }
-}
-
-private struct StatusBadge: View {
-    let status: CloudBackupWalletStatus
-
-    private var label: String {
-        switch status {
-        case .dirty: "Dirty"
-        case .uploading: "Uploading"
-        case .uploadedPendingConfirmation: "Uploaded, confirming"
-        case .confirmed: "Confirmed"
-        case .failed: "Failed"
-        case .deletedFromDevice: "Not on device"
-        case .unsupportedVersion: "Unsupported"
-        case .remoteStateUnknown: "Unknown"
-        }
-    }
-
-    private var color: Color {
-        switch status {
-        case .dirty: .statusWarning
-        case .uploading, .uploadedPendingConfirmation: .statusInfo
-        case .confirmed: .statusSuccess
-        case .failed: .statusError
-        case .deletedFromDevice, .unsupportedVersion: .statusWarning
-        case .remoteStateUnknown: .secondary
-        }
-    }
-
-    var body: some View {
-        Text(label)
-            .font(.caption)
-            .fontWeight(.medium)
-            .foregroundColor(color)
-            .padding(.horizontal, 8)
-            .padding(.vertical, 3)
-            .background(color.opacity(0.15), in: Capsule())
-    }
-}
-
-private struct GroupKey: Hashable, Comparable {
-    let network: Network?
-    let walletMode: WalletMode?
-
-    var title: String {
-        guard let network, let walletMode else {
-            return "Unsupported"
-        }
-
-        return switch walletMode {
-        case .decoy: "\(network.displayName()) · Decoy"
-        default: network.displayName()
-        }
-    }
-
-    static func < (lhs: GroupKey, rhs: GroupKey) -> Bool {
-        if lhs.network == nil || lhs.walletMode == nil {
-            return rhs.network != nil && rhs.walletMode != nil
-        }
-        if rhs.network == nil || rhs.walletMode == nil {
-            return false
-        }
-
-        let lhsNetwork = lhs.network!
-        let rhsNetwork = rhs.network!
-        if lhsNetwork != rhsNetwork {
-            return lhsNetwork.displayName() < rhsNetwork.displayName()
-        }
-        return lhs.walletMode == .main && rhs.walletMode != .main
     }
 }
